Util functions moved to a separate file
