+++ benchmark/results/column_cache.txt
@@ -1,0 +1,1 @@
+0 18 5 2 10 1 2110554 99010 952381 952381 909091 465116 476190 370370 232558 134228 73801 40241 17969 7452 4884 3460 3267 1670 850 442 0 17746 909091 952381 476190 500000 363636 208333 116959 61162 27933 16353 7713 4317 4571 2242 1098 560 289 0 0 42553 800000 1000000 740741 500000 259740 153846 83333 39920 20305 10695 5313 2670 1342 660 333 167 0 0 0 35842 714286 500000 294118 166667 90090 47281 22247 11142 5940 2960 1473 733 370 182 94 0 0 0 0 28209 246914 150376 90090 47847 24450 11955 5466 3106 1561 776 383 193 97 49 0 0 0 0 0 20986 43103 32362 17212 8741 4223 2636 1544 805 402 201 100 50 25 0 0 0 0 0 0 19102 24631 12634 6517 3209 1583 817 408 203 102 51 25 13 0 0 0 0 0 0 0 9416 6502 3263 1632 800 404 206 102 51 26 13 6 0 0 0 0 0 0 0 0 2989 1644 822 406 206 103 51 26 13 6 5 0 0 0 0 0 0 0 0 0 803 412 205 103 52 26 13 6 5 5 0 0 0 0 0 0 0 0 0 0 205 103 52 26 13 6 5 5 4 0 0 0 0 0 0 0 0 0 0 0 51 26 13 6 5 5 4 2 0 0 0 0 0 0 0 0 0 0 0 0 13 6 5 5 4 2 1 0 0 0 0 0 0 0 0 0 0 0 0 0 5 5 4 2 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 4 2 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 1 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0.0202 0.0021 0.0021 0.0022 0.0043 0.0042 0.0054 0.0086 0.0149 0.0271 0.0497 0.1113 0.2684 0.4095 0.578 0.6122 1.1973 2.3529 4.5219 0 0.1127 0.0022 0.0021 0.0042 0.004 0.0055 0.0096 0.0171 0.0327 0.0716 0.1223 0.2593 0.4633 0.4375 0.8919 1.8213 3.5735 6.9264 0 0 0.047 0.0025 0.002 0.0027 0.004 0.0077 0.013 0.024 0.0501 0.0985 0.187 0.3764 0.7492 1.4906 3.0299 6.0146 11.9477 0 0 0 0.0558 0.0028 0.004 0.0068 0.012 0.0222 0.0423 0.0899 0.1795 0.3367 0.6756 1.358 2.7294 5.4101 10.9674 21.3573 0 0 0 0 0.0709 0.0081 0.0133 0.0222 0.0418 0.0818 0.1673 0.3659 0.644 1.281 2.5785 5.2225 10.3533 20.5305 41.1142 0 0 0 0 0 0.0953 0.0464 0.0618 0.1162 0.2288 0.4736 0.7587 1.2954 2.4833 4.976 9.9681 19.9763 39.9452 79.4863 0 0 0 0 0 0 0.1047 0.0812 0.1583 0.3069 0.6233 1.2634 2.4474 4.9006 9.8349 19.6285 39.2619 79.1037 157.597 0 0 0 0 0 0 0 0.2124 0.3076 0.613 1.2258 2.4996 4.9494 9.7291 19.5212 38.9443 78.2715 155.775 312.472 0 0 0 0 0 0 0 0 0.6691 1.2165 2.4328 4.9277 9.6886 19.3968 39.1684 77.4888 155.277 311.055 626.354 0 0 0 0 0 0 0 0 0 2.4899 4.8525 9.7525 19.3562 38.7219 77.6056 154.992 309.501 622.165 1246.11 0 0 0 0 0 0 0 0 0 0 9.7431 19.3942 38.6877 77.4882 154.535 309.578 622.381 1251.21 2560.56 0 0 0 0 0 0 0 0 0 0 0 39.115 77.197 154.679 310.27 618.284 1238.01 2489.25 4984.25 0 0 0 0 0 0 0 0 0 0 0 0 154.869 309.891 619.371 1237.11 2477.97 5006.84 10023.6 0 0 0 0 0 0 0 0 0 0 0 0 0 619.856 1241.42 2512.68 5199.45 10458.2 20089.8 0 0 0 0 0 0 0 0 0 0 0 0 0 0 2490.79 4948.87 9940.35 20036 39831.7 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 9953.11 19945 39829.4 80254 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 41131.2 80033.5 159861 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 161482 320329 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 639882
