+++ benchmark/results/fft_cache.txt
@@ -1,0 +1,1 @@
+0 18 5 2 10 1 405651 110497 555556 384615 238095 125786 61538 21858 13793 5804 2870 1339 622 287 134 63 30 14 7 5 0 40241 350877 219780 118343 59172 28736 13587 5528 2884 1351 622 289 135 64 30 14 7 5 0 0 22297 101523 119048 59701 29197 13680 5543 2884 1345 620 287 135 64 30 14 7 5 0 0 0 36630 116959 60976 29112 13671 5454 2882 1339 620 288 135 63 30 14 7 5 0 0 0 0 30912 58824 28860 13624 5519 2875 1346 616 290 136 64 30 14 7 5 0 0 0 0 0 24390 28409 13578 5488 2880 1328 617 290 136 64 30 14 7 5 0 0 0 0 0 0 17109 13596 5274 2921 1341 615 288 135 64 30 14 7 5 0 0 0 0 0 0 0 9794 5281 2902 1335 623 289 135 64 30 14 7 5 0 0 0 0 0 0 0 0 4666 2913 1345 620 289 135 63 28 14 7 5 0 0 0 0 0 0 0 0 0 2492 1334 619 288 135 64 30 14 7 5 0 0 0 0 0 0 0 0 0 0 1232 620 287 136 63 30 14 7 5 0 0 0 0 0 0 0 0 0 0 0 593 286 135 63 30 14 7 5 0 0 0 0 0 0 0 0 0 0 0 0 283 135 63 30 14 7 5 0 0 0 0 0 0 0 0 0 0 0 0 0 134 63 30 14 7 5 0 0 0 0 0 0 0 0 0 0 0 0 0 0 63 30 14 7 5 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 30 14 7 5 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 14 7 5 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 7 5 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 5 0.0181 0.0036 0.0052 0.0084 0.0159 0.0325 0.0915 0.145 0.3446 0.6969 1.4938 3.2179 6.9657 14.8917 31.6269 67.1874 141.682 299.864 652.017 0 0.0497 0.0057 0.0091 0.0169 0.0338 0.0696 0.1472 0.3618 0.6935 1.4807 3.214 6.9242 14.7861 31.385 66.8327 141.641 299.901 651.664 0 0 0.0897 0.0197 0.0168 0.0335 0.0685 0.1462 0.3608 0.6936 1.4873 3.2239 6.9733 14.8173 31.3154 66.997 141.555 302.414 651.385 0 0 0 0.0546 0.0171 0.0328 0.0687 0.1463 0.3667 0.694 1.4938 3.2246 6.945 14.8006 31.5052 66.4635 144.824 299.811 651.621 0 0 0 0 0.0647 0.034 0.0693 0.1468 0.3624 0.6957 1.4858 3.248 6.8911 14.7072 31.3303 66.6041 141.711 299.821 650.79 0 0 0 0 0 0.082 0.0704 0.1473 0.3644 0.6945 1.5055 3.2396 6.9024 14.7519 31.372 66.4962 141.806 299.59 651.325 0 0 0 0 0 0 0.1169 0.1471 0.3792 0.6846 1.4909 3.2505 6.9451 14.7926 31.4545 66.8431 141.867 302.532 651.76 0 0 0 0 0 0 0 0.2042 0.3787 0.6891 1.4977 3.2116 6.9297 14.7857 31.4939 66.7272 141.642 299.986 650.604 0 0 0 0 0 0 0 0 0.4286 0.6866 1.4869 3.2249 6.9116 14.8185 31.507 70.7028 141.882 300.447 651.278 0 0 0 0 0 0 0 0 0 0.8025 1.4993 3.2319 6.9491 14.8179 31.4897 66.7178 141.424 300.086 654.649 0 0 0 0 0 0 0 0 0 0 1.6237 3.2235 6.9709 14.7601 31.5618 66.4886 142.384 299.617 651.52 0 0 0 0 0 0 0 0 0 0 0 3.3728 7.0019 14.8246 31.5505 66.4519 142.16 302.877 650.99 0 0 0 0 0 0 0 0 0 0 0 0 7.0696 14.8556 31.6488 66.5033 141.983 300.032 652.184 0 0 0 0 0 0 0 0 0 0 0 0 0 14.9428 31.6442 66.6909 142.326 300.199 651.394 0 0 0 0 0 0 0 0 0 0 0 0 0 0 31.7451 67.4561 141.972 300.018 654.155 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 66.8891 141.936 299.785 652.291 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 142.829 300.059 650.994 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 301.548 653.625 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 655.035
